+++ src/types.jl
@@ -75,4 +75,3 @@ struct Topography <: OA
     metadata::Dict{String,Any}
     data::Matrix{Float64}
 end
-